requirements.txt
JPype1>=1.0.1
tox==3.20.1
pytest==6.1.1
pytest-cov==2.10.1
coveralls==2.1.2
sphinx_rtd_theme==0.5.0